Update nn-from-scratch.ipynb
Found a typo in the backpropagation formula in the ipython notebook file on line 290. It should be 1 - tanh^2(z1), rather than 1-tanh^2(z2). It is correctly implemented in the code below and even in the description in the blog post: http://www.wildml.com/2015/09/implementing-a-neural-network-from-scratch/. 

Also added subscripts to delta_2 on lines 293 and 294.